final commit of code ! added all of the fucking stuff and made the code more beutifull
