+++ src/pepse/transitions/HorizontalTransitionCreator.java
@@ -2,16 +2,15 @@ package pepse.transitions;
 
 import danogl.GameObject;
 import danogl.components.Transition;
-import pepse.util.TransitionCreator;
 
 /**
- * horizontal movement transition executor
+ * wooboling horizontal movement transition executor
  *
  * @author Ruth Yukhnovetsky
  */
 public class HorizontalTransitionCreator implements TransitionCreator {
 
-    private final float finalValue;
+    private final float horizontalSpeed;
     private final Transition.Interpolator<Float> interpolator;
     private final Transition.TransitionType transitionType;
     private final Runnable onTransitionFinishedCallback;
@@ -19,29 +18,36 @@ public class HorizontalTransitionCreator implements TransitionCreator {
     /**
      * constructor
      *
-     * @param finalValue                   final location of horizontal movement
+     * @param horizontalSpeed              max speed of the horizontal movement
      * @param interpolator                 interpolator
      * @param transitionType               type of transition
      * @param onTransitionFinishedCallback callback
      */
-    public HorizontalTransitionCreator(float finalValue,
+    public HorizontalTransitionCreator(float horizontalSpeed,
                                        Transition.Interpolator<Float> interpolator,
                                        Transition.TransitionType transitionType,
                                        Runnable onTransitionFinishedCallback) {
 
-        this.finalValue = finalValue;
+        this.horizontalSpeed = horizontalSpeed;
         this.interpolator = interpolator;
         this.transitionType = transitionType;
         this.onTransitionFinishedCallback = onTransitionFinishedCallback;
     }
 
+    /**
+     * Creates a horizontal movement transition
+     *
+     * @param cycleLength the transition cycle length
+     * @param gameObject  the game object that transition wil be executed on
+     * @return created transitions
+     */
     @Override
     public Transition[] createTransitions(float cycleLength, GameObject gameObject) {
         var horizontalTransition = new Transition<>(
                 gameObject,
                 gameObject.transform()::setVelocityX,
-                this.finalValue * -1,
-                this.finalValue,
+                this.horizontalSpeed * -1,
+                this.horizontalSpeed,
                 this.interpolator,
                 cycleLength,
                 this.transitionType,
