+++ src/pepse/util/MovementOptions.java
@@ -1,5 +1,8 @@
 package pepse.util;
 
+/**
+ * Enum for different options of movement in the simulator
+ */
 public enum MovementOptions {
     Standing,
     Flying,
