+++ src/pepse/world/Terrain.java
@@ -5,13 +5,15 @@ import danogl.gui.rendering.RectangleRenderable;
 import danogl.util.Vector2;
 import pepse.util.ColorSupplier;
 import pepse.util.GroundHeightCalculator;
-import pepse.util.PerlinNoise;
-import pepse.util.SurfaceCreator;
+import pepse.util.surface.PerlinNoise;
+import pepse.util.surface.SurfaceCreator;
 
 import java.awt.*;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+
+import static pepse.configuration.GameObjectsConfiguration.BLOCK_SIZE;
 
 /**
  * creating necessary blocks and lets other objects know the height of the terrain at a certain coordinate.
@@ -71,7 +73,7 @@ public class Terrain implements GroundHeightCalculator, SurfaceCreator {
         if (groundHeights.containsKey((int) x))
             return groundHeights.get((int) x);
         var value = INITAL_GROUND_LEVEL +
-                (float) this.noiseCreator.noise(x) * NOISE_FACTOR * Block.SIZE;
+                (float) this.noiseCreator.noise(x) * NOISE_FACTOR * BLOCK_SIZE;
         groundHeights.put((int) x, value);
         return value;
     }
@@ -85,7 +87,7 @@ public class Terrain implements GroundHeightCalculator, SurfaceCreator {
      */
     @Override
     public float gameObjectHeightAt(float x) {
-        return (float) Math.floor(groundHeightAt(x) / Block.SIZE) * Block.SIZE;
+        return (float) Math.floor(groundHeightAt(x) / BLOCK_SIZE) * BLOCK_SIZE;
     }
 
 
@@ -111,11 +113,11 @@ public class Terrain implements GroundHeightCalculator, SurfaceCreator {
         for (var i = 0; i < TERRAIN_DEPTH; i++) {
             if (i < 2) {
                 blocks.add(new ReactingBlock(
-                        new Vector2(x, this.windowDimensions.y() - (y - (Block.SIZE * i))),
+                        new Vector2(x, this.windowDimensions.y() - (y - (BLOCK_SIZE * i))),
                         new RectangleRenderable(ColorSupplier.approximateColor(BASE_GROUND_COLOR))));
             } else {
                 blocks.add(new Block(
-                        new Vector2(x, this.windowDimensions.y() - (y - (Block.SIZE * i))),
+                        new Vector2(x, this.windowDimensions.y() - (y - (BLOCK_SIZE * i))),
                         new RectangleRenderable(ColorSupplier.approximateColor(BASE_GROUND_COLOR))));
             }
         }
@@ -133,7 +135,7 @@ public class Terrain implements GroundHeightCalculator, SurfaceCreator {
         var options = new ArrayList<Integer>();
         var realFrom = getClosestToBlockSize(fromRange);
         var realTo = getClosestToBlockSize(toRange);
-        for (var x = realFrom; x <= realTo; x += Block.SIZE) {
+        for (var x = realFrom; x <= realTo; x += BLOCK_SIZE) {
             options.add(x);
         }
         return options;
@@ -142,7 +144,7 @@ public class Terrain implements GroundHeightCalculator, SurfaceCreator {
 
     private int getClosestToBlockSize(int x) {
         //calculate the real 'range' to move on (as a result of Block size)
-        return x - x % Block.SIZE;
+        return x - x % BLOCK_SIZE;
     }
 
 }
