+++ src/pepse/world/daynight/SunHalo.java
@@ -5,6 +5,7 @@ import danogl.collisions.GameObjectCollection;
 import danogl.components.CoordinateSpace;
 import danogl.gui.rendering.OvalRenderable;
 import danogl.util.Vector2;
+import pepse.configuration.GameObjectsConfiguration;
 
 import java.awt.*;
 
@@ -14,7 +15,7 @@ import java.awt.*;
  */
 public class SunHalo {
 
-    private static int SUN_HALO_SIZE = 210;
+
 
     /**
      * creating sun halo
@@ -32,7 +33,7 @@ public class SunHalo {
 
         var haloRenderable = new OvalRenderable(color);
         var sunHalo = new GameObject(Vector2.ZERO,
-                new Vector2(SUN_HALO_SIZE, SUN_HALO_SIZE),
+                GameObjectsConfiguration.SUN_HALO_SIZE,
                 haloRenderable);
         sunHalo.setCoordinateSpace(CoordinateSpace.CAMERA_COORDINATES);
         sunHalo.setCenter(sun.getCenter());
