+++ src/pepse/world/movement/MovementAnimation.java
@@ -1,6 +1,6 @@
 package pepse.world.movement;
 
-import pepse.util.AnimatedGameObject;
+import pepse.util.unique_game_objects.AnimatedGameObject;
 
 /**
  * interface for executing render change and animations
