+++ src/pepse/world/trees/leaf/LeafLifeDeathCycle.java
@@ -5,7 +5,7 @@ import danogl.components.ScheduledTask;
 import pepse.configuration.GameLayers;
 import pepse.configuration.GameObjectsConfiguration;
 import pepse.configuration.TransitionConfiguration;
-import pepse.util.TransitionCreator;
+import pepse.transitions.TransitionCreator;
 
 import java.util.Random;
 
@@ -17,7 +17,7 @@ public class LeafLifeDeathCycle {
     private final TransitionCreator leafFallingTransformation;
     private final GameObjectCollection collection;
     private final Random random;
-    private final static int GRAVITY = 70;
+    private final static int LEAF_GRAVITY = 70;
 
     /**
      * constructor
@@ -47,23 +47,34 @@ public class LeafLifeDeathCycle {
                 () -> executeLeafFall(leaf));
     }
 
+    /**
+     * executing the leaf falling
+     *
+     * @param leaf leaf to fall
+     */
     private void executeLeafFall(Leaf leaf) {
-        //replace original leaf with falling leaf:
+        // replace original leaf with falling leaf
+        // this is because of performance ! we only check collision of the falling trees!
         collection.removeGameObject(leaf, GameLayers.LEAF_LAYER);
         var fallingLeaf = new FallingLeaf(leaf, this.leafFallingTransformation);
 
-        fallingLeaf.transform().setAccelerationY(GRAVITY);
+        fallingLeaf.transform().setAccelerationY(LEAF_GRAVITY);
         collection.addGameObject(fallingLeaf, GameLayers.LEAF_LAYER);
         //when fadeout finishes, remove the leaf that fell and replace with a new one on tree
         fallingLeaf.renderer().fadeOut(
                 TransitionConfiguration.LEAF_FADEOUT,
-                () -> {collection.removeGameObject(fallingLeaf, GameLayers.LEAF_LAYER);
-                       collection.addGameObject(leaf, GameLayers.LEAF_LAYER);
+                () -> {
+                    collection.removeGameObject(fallingLeaf, GameLayers.LEAF_LAYER);
+                    collection.addGameObject(leaf, GameLayers.LEAF_LAYER);
                 });
     }
 
+    /**
+     * Generates a random lifetime length for a given tree
+     *
+     * @return life time length
+     */
     private float getLeafLifeLength() {
-        //gets a random lifetime length for a leaf
         return this.random.nextInt(GameObjectsConfiguration.LEAF_MAX_LIFE -
                 GameObjectsConfiguration.LEAF_MIN_LIFE) + GameObjectsConfiguration.LEAF_MIN_LIFE;
     }
