+++ src/pepse/world/trees/leaf/LeafTransitionHandler.java
@@ -2,7 +2,7 @@ package pepse.world.trees.leaf;
 
 import danogl.components.ScheduledTask;
 import pepse.configuration.TransitionConfiguration;
-import pepse.util.TransitionCreator;
+import pepse.transitions.TransitionCreator;
 
 import java.util.Random;
 
@@ -18,7 +18,6 @@ public class LeafTransitionHandler {
     private final Random random;
 
     private final static int OPACITY_MIN_CHANCE = 3;
-
     private final static int RANDOM_NUMBER_RANGE = 5;
 
     /**
@@ -49,13 +48,15 @@ public class LeafTransitionHandler {
         float waitTime = TransitionConfiguration.LEAF_DELAYS[
                 this.random.nextInt(TransitionConfiguration.LEAF_DELAYS.length)];
 
-        if (randomNumber == 1) { //activating angle and size transaction
+        if (randomNumber == 1) {
+            //activating angle and size transaction
             new ScheduledTask(leaf,
                     waitTime,
                     repeat,
                     () -> this.angleAndSizeTransition.createTransitions(cycleLength, leaf));
         }
-        if (randomNumber <= OPACITY_MIN_CHANCE) { //activating opacity changer transaction
+        if (randomNumber <= OPACITY_MIN_CHANCE) {
+            //activating opacity changer transaction
             this.opacityTransition.createTransitions(cycleLength - waitTime, leaf);
         }
     }
